Home Page added and modified
