+++ src/App.js
@@ -1,21 +1,20 @@
 import React from "react";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Layout from "./hocs/Layout";
-// import Home from "./pages/Home";
-import Home from './component/Home';
-import "./App.css"
+import Home from "./pages/Home";
+import "./App.css";
 
 function App() {
   return (
     <div className="App">
-    <Router>
-      <Layout>
-        <Switch>
-          <Route exact path="/" component={Home} />
-        </Switch>
-      </Layout>
-    </Router>
-</div>
+      <Router>
+        <Layout>
+          <Switch>
+            <Route exact path="/" component={Home} />
+          </Switch>
+        </Layout>
+      </Router>
+    </div>
   );
 }
