+++ src/style/Home.css
@@ -9,92 +9,17 @@
     justify-content: center;
     align-items: center;
 }
-.img_slider{
-    width: 100%;
+  .carousel .carousel-item {
     height: 500px;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-.slider{
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-    position: relative;
-}
-.slides{
-    height: 500px;
-    object-fit: cover;
-    background-size: cover;
-    position: relative;
-    transition: 900ms cubic-bezier(0.165, 0.84, 0.44, 1);
-}
-.slide-img{
-    position: absolute;
-    height: 100%;
-    width: 100%;
-}
+    width: 100vw;
+  }
 
-.slide-img img{
-    background-repeat: no-repeat;
-    object-fit:cover;
-    background-size: cover;
-    width: 100%;
-    height: 100%;
-}
-.left,
-.right{
-    background: rgba(0, 0, 0, 0.4);
-    padding: 15px;
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    z-index: 1000;
-    cursor: pointer;
-    font-size: 26px;
-    color: white;
-    transition: 500ms;
-}
-.left:hover{
-    background: rgb(71, 71, 71,0.849);
-}
-.right:hover{
-    background: rgba(71, 71, 71, 0.849);
-}
-.left{
-    left: 0;
-}
-.right{
-    right: 0;
-}
-.lines{
-    display: flex;
-    height: 32px;
-    width: 100%;
-    align-items: center;
-    justify-content:center;
-    margin: 16px 0;
-    display: none;
-}
-.li{
-    background: darkgray;
-    height: 6px ;
-    width: 5%;
-    /* border: 2px solid #000; */
-    cursor: pointer;
-    transition: 400ms;
-    margin: 0 8px;
-}
-.li.active{
-    background: rgb(32, 32, 32);
-    
-}
-.none{
-    display: none;
-}
-
+  
+/* slider img  */
+.carousel-item img {
+      width: 100vw;
+      height: 500px;
+  }
 .books_to_baught{
     display: flex;
     flex-direction: column;
@@ -103,6 +28,10 @@
     width: 100%;
     margin-top: 10px;
 }
+
+
+
+
 .books_to_baught_head{
     width: 100%;
     display: flex;
